HADOOP-16520. race condition in DDB table init and waiting threads - compile and verify OK after refactor
